Add NewAssetProviderFromLocator function
Allows plugins to reference resolved assets. Refactors some plugins
package functions to pluginsinternal to do this.
