+++ framework/internal/pathsinternal/paths.go
