+++ framework/plugins/plugin.go
@@ -17,7 +17,6 @@ package plugins
 import (
 	"fmt"
 	"io"
-	"os"
 	"path"
 	"sort"
 	"strings"
@@ -27,12 +26,9 @@ import (
 	"github.com/palantir/godel/framework/artifactresolver"
 	"github.com/palantir/godel/framework/godellauncher"
 	"github.com/palantir/godel/framework/internal/pathsinternal"
+	"github.com/palantir/godel/framework/internal/pluginsinternal"
 	"github.com/palantir/godel/framework/pluginapi/v2/pluginapi"
 	"github.com/palantir/godel/pkg/osarch"
-)
-
-const (
-	indentSpaces = 4
 )
 
 // pluginInfoWithAssets bundles a pluginapi.Info with the locators of all the assets specified for it.
@@ -68,7 +64,7 @@ func LoadPluginsTasks(pluginsParam godellauncher.PluginsParam, stdout io.Writer)
 	for k := range plugins {
 		sortedPluginLocators = append(sortedPluginLocators, k)
 	}
-	sortLocators(sortedPluginLocators)
+	pluginsinternal.SortLocators(sortedPluginLocators)
 
 	var tasks []godellauncher.Task
 	var upgradeConfigTasks []godellauncher.UpgradeConfigTask
@@ -114,7 +110,7 @@ func resolvePlugins(pluginsDir, assetsDir, downloadsDir string, osArch osarch.OS
 	plugins := make(map[artifactresolver.Locator]pluginInfoWithAssets)
 	pluginErrors := make(map[artifactresolver.Locator]error)
 	for _, currPlugin := range pluginsParam.Plugins {
-		currPluginLocator, ok := resolveAndVerify(
+		currPluginLocator, ok := pluginsinternal.ResolveAndVerify(
 			currPlugin.LocatorWithResolverParam,
 			pluginErrors,
 			pluginsDir,
@@ -133,7 +129,7 @@ func resolvePlugins(pluginsDir, assetsDir, downloadsDir string, osArch osarch.OS
 		}
 
 		// plugin has been successfully resolved: resolve assets for plugin
-		assetInfoMap, err := resolveAssets(assetsDir, downloadsDir, currPlugin.Assets, osArch, pluginsParam, stdout)
+		assetInfoMap, err := pluginsinternal.ResolveAssets(assetsDir, downloadsDir, currPlugin.Assets, osArch, pluginsParam.DefaultResolvers, stdout)
 		if err != nil {
 			pluginErrors[currPluginLocator] = errors.Wrapf(err, "failed to get asset(s) for plugin %+v", currPluginLocator)
 			continue
@@ -154,113 +150,13 @@ func resolvePlugins(pluginsDir, assetsDir, downloadsDir string, osArch osarch.OS
 	for k := range pluginErrors {
 		sortedKeys = append(sortedKeys, k)
 	}
-	sortLocators(sortedKeys)
+	pluginsinternal.SortLocators(sortedKeys)
 
 	errStringsParts := []string{fmt.Sprintf("failed to resolve %d plugin(s):", len(pluginErrors))}
 	for _, k := range sortedKeys {
 		errStringsParts = append(errStringsParts, pluginErrors[k].Error())
 	}
-	return nil, errors.New(strings.Join(errStringsParts, "\n"+strings.Repeat(" ", indentSpaces)))
-}
-
-func resolveAssets(assetsDir, downloadsDir string, assetParams []artifactresolver.LocatorWithResolverParam, osArch osarch.OSArch, pluginsParam godellauncher.PluginsParam, stdout io.Writer) ([]artifactresolver.Locator, error) {
-	if len(assetParams) == 0 {
-		return nil, nil
-	}
-
-	var assets []artifactresolver.Locator
-	assetErrors := make(map[artifactresolver.Locator]error)
-	for _, currAsset := range assetParams {
-		currAssetLocator, ok := resolveAndVerify(
-			currAsset,
-			assetErrors,
-			assetsDir,
-			downloadsDir,
-			pluginsParam.DefaultResolvers,
-			osArch,
-			stdout,
-		)
-		if !ok {
-			continue
-		}
-		assets = append(assets, currAssetLocator)
-	}
-	sortLocators(assets)
-
-	if len(assetErrors) == 0 {
-		return assets, nil
-	}
-
-	// encountered errors: summarize and return
-	var sortedKeys []artifactresolver.Locator
-	for k := range assetErrors {
-		sortedKeys = append(sortedKeys, k)
-	}
-	sortLocators(sortedKeys)
-
-	errStringsParts := []string{fmt.Sprintf("failed to resolve %d asset(s):", len(assetErrors))}
-	for _, k := range sortedKeys {
-		errStringsParts = append(errStringsParts, assetErrors[k].Error())
-	}
-	return nil, errors.New(strings.Join(errStringsParts, "\n"+strings.Repeat(" ", indentSpaces)))
-}
-
-func resolveAndVerify(
-	currArtifact artifactresolver.LocatorWithResolverParam,
-	artifactErrors map[artifactresolver.Locator]error,
-	dstBaseDir, downloadsDir string,
-	defaultResolvers []artifactresolver.Resolver,
-	osArch osarch.OSArch,
-	stdout io.Writer) (currLocator artifactresolver.Locator, ok bool) {
-
-	currLocator = currArtifact.LocatorWithChecksums.Locator
-	currDstPath := path.Join(dstBaseDir, pathsinternal.PluginFileName(currLocator))
-
-	if _, err := os.Stat(currDstPath); os.IsNotExist(err) {
-		tgzDstPath := path.Join(downloadsDir, pathsinternal.PluginFileName(currLocator)+".tgz")
-		if err := artifactresolver.ResolveArtifactTGZ(currArtifact, defaultResolvers, osArch, tgzDstPath, stdout); err != nil {
-			artifactErrors[currLocator] = err
-			return currLocator, false
-		}
-
-		if err := func() (rErr error) {
-			pluginFile, err := os.OpenFile(currDstPath, os.O_RDWR|os.O_CREATE|os.O_TRUNC, 0755)
-			if err != nil {
-				return errors.Wrapf(err, "failed to create file %s", currDstPath)
-			}
-			defer func() {
-				if err := pluginFile.Close(); err != nil {
-					rErr = errors.Wrapf(err, "failed to close file %s", currDstPath)
-				}
-			}()
-
-			tgzFile, err := os.Open(tgzDstPath)
-			if err != nil {
-				return errors.Wrapf(err, "failed to open %s for reading", tgzDstPath)
-			}
-
-			if err := artifactresolver.CopySingleFileTGZContent(pluginFile, tgzFile); err != nil {
-				return err
-			}
-			return nil
-		}(); err != nil {
-			artifactErrors[currLocator] = errors.Wrapf(err, "failed to extract artifact from archive into destination")
-			return currLocator, false
-		}
-	}
-
-	if wantChecksum, ok := currArtifact.LocatorWithChecksums.Checksums[osArch]; ok {
-		gotChecksum, err := artifactresolver.SHA256ChecksumFile(currDstPath)
-		if err != nil {
-			artifactErrors[currLocator] = errors.Wrapf(err, "failed to compute checksum for plugin")
-			return currLocator, false
-		}
-		if gotChecksum != wantChecksum {
-			artifactErrors[currLocator] = errors.Errorf("failed to verify checksum for %s: want %s, got %s", currDstPath, wantChecksum, gotChecksum)
-			return currLocator, false
-		}
-	}
-	return currLocator, true
+	return nil, errors.New(strings.Join(errStringsParts, "\n"+strings.Repeat(" ", pluginsinternal.IndentSpaces)))
 }
 
 // Verifies that the plugins in the provided map are compatible with one another. Specifically, ensures that:
@@ -287,20 +183,20 @@ func verifyPluginCompatibility(plugins map[artifactresolver.Locator]pluginInfoWi
 	for k := range conflicts {
 		sortedOuterKeys = append(sortedOuterKeys, k)
 	}
-	sortLocators(sortedOuterKeys)
+	pluginsinternal.SortLocators(sortedOuterKeys)
 
 	errString := fmt.Sprintf("%d plugins had compatibility issues:", len(conflicts))
 	for _, k := range sortedOuterKeys {
-		errString += fmt.Sprintf("\n%s%s:", strings.Repeat(" ", indentSpaces), k.String())
+		errString += fmt.Sprintf("\n%s%s:", strings.Repeat(" ", pluginsinternal.IndentSpaces), k.String())
 
 		var sortedInnerKeys []artifactresolver.Locator
 		for innerK := range conflicts[k] {
 			sortedInnerKeys = append(sortedInnerKeys, innerK)
 		}
-		sortLocators(sortedInnerKeys)
+		pluginsinternal.SortLocators(sortedInnerKeys)
 
 		for _, innerK := range sortedInnerKeys {
-			errString += fmt.Sprintf("\n%s%s", strings.Repeat(" ", indentSpaces*2), conflicts[k][innerK].Error())
+			errString += fmt.Sprintf("\n%s%s", strings.Repeat(" ", pluginsinternal.IndentSpaces*2), conflicts[k][innerK].Error())
 		}
 	}
 	return errors.New(errString)
@@ -351,10 +247,4 @@ func verifySinglePluginCompatibility(plugin artifactresolver.Locator, plugins ma
 		}
 	}
 	return errs
-}
-
-func sortLocators(locs []artifactresolver.Locator) {
-	sort.Slice(locs, func(i, j int) bool {
-		return locs[i].String() < locs[j].String()
-	})
 }
